pyproject.toml
[project]
name = "naismith-nerds"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Jason Gardner", email = "jmgjasongardner@gmail.com"},
]
dependencies = [
    "polars>=1.20.0",
    "black>=24.10.0",
    "scikit-learn>=1.6.1",
    "argparse>=1.4.0",
    "pandas>=2.2.3",
    "openpyxl>=3.1.5",
    "pyarrow>=19.0.0",
    "flask>=3.1.0",
    "ngrok>=1.4.0",
    "argparse==1.4.0",
    "black==25.1.0",
    "blinker==1.9.0",
    "click==8.1.8",
    "colorama==0.4.6; platform_system == \"Windows\"",
    "et-xmlfile==2.0.0",
    "flask==3.1.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.5",
    "joblib==1.4.2",
    "markupsafe==3.0.2",
    "mypy-extensions==1.0.0",
    "ngrok==1.4.0",
    "numpy==2.2.2",
    "openpyxl==3.1.5",
    "packaging==24.2",
    "pandas==2.2.3",
    "pathspec==0.12.1",
    "platformdirs==4.3.6",
    "polars==1.22.0",
    "pyarrow==19.0.0",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.1",
    "scipy==1.15.1",
    "six==1.17.0",
    "threadpoolctl==3.5.0",
    "tzdata==2025.1",
    "werkzeug==3.1.3",
    "gunicorn>=23.0.0",
    "gevent>=24.11.1",
    "psutil>=7.0.0",
    "duckdb>=1.2.1",
    "plotly>=6.0.1",
    "pillow>=12.0.0",
    "msal>=1.31.0",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
